修复统一应答时私吞业务异常的bug 修复之前规定的日期时间格式数据不支持以url编码方式传参的get请求的bug 优化返回的体温信息封装数据

+++ resident-temperature-monitoring-system-backstage-service/resident-temperature-monitoring-system-backstage-dao/src/main/java/com/lwh147/rtms/backstage/dao/entity/Resident.java
@@ -53,7 +53,7 @@ public class Resident {
      */
     @Column(name = "room")
     @ApiModelProperty("房间号")
-    private Byte room;
+    private Integer room;
 
     /**
      * 联系电话
