+++ resident-temperature-monitoring-system-backstage-service/resident-temperature-monitoring-system-backstage-pojo/src/main/java/com/lwh147/rtms/backstage/pojo/dto/ResidentDTO.java
@@ -45,7 +45,7 @@ public class ResidentDTO {
      * 房间号
      */
     @ApiModelProperty("房间号")
-    private Byte room;
+    private Integer room;
 
     /**
      * 联系电话
